LEVEL 1 + LEVEL 2 empezando

+++ starter_code/game.js
@@ -2,7 +2,7 @@ function Game(canvadId) {
     this.canvas = document.getElementById(canvadId);
     this.ctx = this.canvas.getContext("2d");
     this.fps = 40;
-
+    this.level=1;
 
     
     this.player1=new Player(this,38,39,40,37,18,16);
@@ -58,6 +58,24 @@ function Game(canvadId) {
   };*/
 
   Game.prototype.nextClient = function() {
+
+    alert('this.player1.score'+this.player1.score);
+    alert('this.player2.score'+this.player2.score);
+
+        if(this.player1.score===3 ||this.player2.score===3){
+
+
+        this.level=2;
+
+        }else if(this.player1.score===6 ||this.player2.score===6){
+
+
+          this.level=3;
+  
+          }
+
+       
+
       this.stop();
       this.reset();
       this.start();
@@ -65,15 +83,18 @@ function Game(canvadId) {
   };
 
   Game.prototype.reset = function() {
+if(this.level===1){
+    this.background = new Background(this,'images/fondo1.png');
 
-    this.background = new Background(this);
-
-    this.order = new Order(this);
+    this.order = new Order(this,this.level);
     
     this.order.getBurger();
 
     this.player1.delete();
     this.player2.delete();
+
+   this.player1.burgerIngredients();
+   this.player2.burgerIngredients();
 
     this.player1.setListeners();
     this.player2.setListeners();
@@ -85,7 +106,33 @@ function Game(canvadId) {
     this.person.getClient();
 
     this.framesCounter = 0;
+}else if(this.level===2){
+alert('level2');
+  this.background = new Background(this,'images/paris.jpg');
 
+  this.order = new Order(this,1);
+  
+  this.order.getBurger();
+
+  this.player1.delete();
+  this.player2.delete();
+
+ this.player1.burgerIngredients();
+ this.player2.burgerIngredients();
+
+  this.player1.setListeners();
+  this.player2.setListeners();
+
+  this.player1.burgerOrder(this.order.selectedBurger);
+  this.player2.burgerOrder(this.order.selectedBurger);
+
+  this.person=new Person(this);
+  this.person.getClient();
+
+  this.framesCounter = 0;
+
+
+}
   };
